requirements.txt
streamlit>=1.35.0
langchain>=1.0.0
langchain-core>=0.3.0
langchain-groq
langchain-openai
langchain-pinecone
pinecone-client
python-dotenv